Add GitHub Actions workflows, update demo_spec.js, and reorganize custom commands
Adds two GitHub Actions workflows, with one setting up OpenSearch and OpenSearch Dashboards from source, and the other setting them up from downloaded snapshots. These workflows run automatically upon pushes or pull requests to main or any dev branches. This commit also updates the demo_spec.js test to properly setup the sample data and incorporates the new Cypress custom commands to improve flakiness. Finally, this commit groups the Cypress custom commands into separate files based on their type of functionality, and adds other minor improvements.
